Move reference tests to subdirs
