+++ test/reference/packed_compound_array/schema.zs
@@ -1,0 +1,12 @@
+package packed_compound_array.schema;
+
+struct PackableStructure
+{
+    uint32 value;
+    string text;
+};
+
+struct PackedArrayOfCompounds
+{
+    packed PackableStructure list[5];
+};
