+++ test/reference/packed_nested_array/test.go
@@ -1,7 +1,7 @@
 package reference
 
 import (
-	"gen/github.com/woven-planet/go-zserio/testdata/reference_modules/core/array"
+	"gen/github.com/woven-planet/go-zserio/testdata/packed_nested_array/schema"
 	"os"
 	"testing"
 
@@ -26,7 +26,7 @@ func TestRoundTrip(t *testing.T) {
 	require.NoError(t, err)
 
 	// When
-	var object array.PackedArrayOfCompounds
+	var object schema.PackedNestedArray
 	require.NoError(t, zserio.Unmarshal(want, &object), "unmarshal")
 	got, err := zserio.Marshal(&object)
 	require.NoError(t, err, "marshal")
@@ -41,17 +41,17 @@ func TestEqual(t *testing.T) {
 	require.NoError(t, err)
 
 	// When
-	var got array.PackedArrayOfCompounds
+	var got schema.PackedNestedArray
 	require.NoError(t, zserio.Unmarshal(bytes, &got))
 
 	// Then
-	want := array.PackedArrayOfCompounds{
-		List: []array.PackableStructure{
-			{Value: 15, Text: "fünfzehn"},
-			{Value: 14, Text: "fourteen"},
-			{Value: 13, Text: "dertien"},
-			{Value: 16, Text: "seksten"},
-			{Value: 0, Text: "nul"},
+	want := schema.PackedNestedArray{
+		List: []schema.PackableNestedStructure{
+			{Value32: 15, Text: "fünfzehn", InnerStructure: schema.InnerStructure{Value64: 15 * 15, Value16: 15 * 2}},
+			{Value32: 14, Text: "fourteen", InnerStructure: schema.InnerStructure{Value64: 14 * 14, Value16: 14 * 2}},
+			{Value32: 13, Text: "dertien", InnerStructure: schema.InnerStructure{Value64: 13 * 13, Value16: 13 * 2}},
+			{Value32: 16, Text: "seksten", InnerStructure: schema.InnerStructure{Value64: 16 * 16, Value16: 16 * 2}},
+			{Value32: 0, Text: "nul", InnerStructure: schema.InnerStructure{Value64: 0, Value16: 0 * 2}},
 		},
 	}
 	assert.Equal(t, want, got)
