--- test/reference/packed_uint32_array.py
@@ -1,6 +1,0 @@
-from testdata.reference_modules.core.array.packed_uint32_array import PackedUint32Array
-
-
-def new():
-    array = PackedUint32Array([15, 14, 13, 16, 0])
-    return array
