+++ test/reference/packed_uint32_array/schema.zs
@@ -1,0 +1,8 @@
+package packed_uint32_array.schema;
+
+
+// Examples taken from https://github.com/ndsev/zserio/blob/master/doc/ZserioInvisibles.md#packed-arrays-of-integers
+struct PackedUint32Array
+{
+    packed uint32 list[5];
+};
