+++ testdata/reference_modules/all.zs
@@ -1,6 +1,5 @@
 package reference_modules.all;
 
-import reference_modules.core.array.*;
 import reference_modules.core.instantiations.*;
 import reference_modules.core.templates.*;
 import reference_modules.core.types.*;
